Create sml_v1
This is an implementation for SML approach mentioned in the article "Local Search Approaches in Stable Matching Problems" and aims to find a stable matching given men's and women's preference lists.
However, this approach does not use random restart or random walk when neighbors' evaluation value is less than the current state's, and it takes a step to a neighbor if neighbor's evaluation value is equal to the current state's evaluation value. Therefore, giving large inputs may cause this algorithm to be trapped in infinite loops.

So, this algorithm is implemented for small input sizes such as 10.
In sml_v2, I apply random restart and in sml_v3, I apply random walk.